src/requirements.txt
Flask
pymysql
cohere
fpdf
pandas
openpyxl
datetime
matplotlib
python-dotenv
reportlab
docx
pymupdf
fitz
psycopg2-binary

